refactor : 아티클 파싱 클래스 분할 - 아티클 파싱 전용 클래스 분할
